+++ react/RecommendationShelf.tsx
@@ -1,12 +1,8 @@
-import React, { Fragment, useMemo } from 'react'
+import React from 'react'
 import { defineMessages } from 'react-intl'
-import { useRuntime } from 'vtex.render-runtime'
-import { useSearchPage } from 'vtex.search-page-context/SearchPageContext'
-import { useProduct } from 'vtex.product-context'
-import { useOrderForm } from 'vtex.order-manager/OrderForm'
 
-import useRecommendation from './hooks/useRecommendation'
-import Shelf from './Shelf'
+import { RecommendationShelfErrorBoundary } from './components/RecommendationShelfErrorBoundary'
+import { RecommendationShelfContainer } from './components/RecommendationShelfContainer'
 
 defineMessages({
   title: {
@@ -16,62 +12,18 @@ defineMessages({
 })
 
 type Props = {
-  campaignId: string
+  campaignVrn?: string
   title?: string
 }
 
 const RecommendationShelf: StorefrontFunctionComponent<Props> = ({
-  campaignId,
+  campaignVrn,
   title,
 }) => {
-  const { searchQuery } = useSearchPage()
-  const productContext = useProduct()
-  const { page } = useRuntime()
-  const {
-    orderForm: { items: orderFormItems },
-  } = useOrderForm()
-
-  let productId: string | undefined
-
-  if (productContext) {
-    const { product } = productContext
-
-    if (product) {
-      productId = product.productId
-    }
-  }
-
-  if (searchQuery) {
-    productId = searchQuery?.products[0]?.productId
-  } else if (orderFormItems && page === 'store.checkout.cart') {
-    productId = orderFormItems?.map((item: any) => item.id)[0]
-  }
-
-  const { data, error } = useRecommendation(campaignId, productId)
-
-  const products = useMemo(() => {
-    if (error || !data) {
-      return undefined
-    }
-
-    const recommended = data.syneriseRecommendationV1.products
-
-    if (recommended && recommended.length > 0) {
-      return recommended
-    }
-
-    return undefined
-  }, [data, error])
-
-  return products?.length ? (
-    <Shelf
-      products={products}
-      title={title}
-      correlationId={data?.syneriseRecommendationV1.correlationId ?? ''}
-      campaignId={campaignId}
-    />
-  ) : (
-    <Fragment />
+  return (
+    <RecommendationShelfErrorBoundary>
+      <RecommendationShelfContainer campaignVrn={campaignVrn} title={title} />
+    </RecommendationShelfErrorBoundary>
   )
 }