+++ react/components/Shelf/index.tsx
@@ -7,7 +7,8 @@ import styles from './styles.css'
 import { notifyClick } from './notifyClick'
 
 type Props = {
-  campaignId: string
+  userId: string
+  campaignVrn: string
   correlationId: string
   products: Product[]
   title?: string
@@ -19,13 +20,14 @@ const Shelf: StorefrontFunctionComponent<Props> = ({
   title,
   products,
   correlationId,
-  campaignId,
+  campaignVrn,
+  userId,
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const onProductClick = (p: Product) => {
     const itemId = p.productId ?? ''
 
-    notifyClick({ itemId, campaignId, correlationId })
+    notifyClick({ productId: itemId, campaignVrn, correlationId, userId })
   }
 
   return (