+++ react/hooks/useRecommendation.tsx
@@ -3,20 +3,25 @@ import { canUseDOM } from 'vtex.render-runtime'
 
 import type { Args, Response } from '../graphql/QueryRecommendationShelf.gql'
 import recommendationQuery from '../graphql/QueryRecommendationShelf.gql'
-import { getCookie } from '../utils/dom'
 
-function useRecommendation(campaignId: string, productId?: string) {
-  const uuid = canUseDOM ? getCookie('_snrs_uuid') : ''
-
+function useRecommendation({
+  campaignVrn,
+  products,
+  userId,
+}: {
+  campaignVrn?: string
+  products: string[]
+  userId: string
+}) {
   const variables = {
-    synUserId: uuid ?? '',
-    campaignId,
-    productId,
+    userId,
+    campaignVrn: campaignVrn ?? '',
+    products,
   }
 
   const { error, data } = useQuery<Response, Args>(recommendationQuery, {
     variables,
-    skip: !canUseDOM || !campaignId || !uuid,
+    skip: !canUseDOM || !campaignVrn || !userId,
     notifyOnNetworkStatusChange: true,
   })