+++ react/utils/dom.ts
@@ -1,6 +1,0 @@
-export function getCookie(name: string): string | null {
-  const regex = new RegExp(`(^|;)[ ]*${name}=([^;]*)`)
-  const match = regex.exec(document.cookie)
-
-  return match ? decodeURIComponent(match[2]) : null
-}